backend/requirements.txt
Django==1.9.7
django-extensions==1.7.4
djangorestframework==3.4.7
psycopg2==2.6.2
six==1.10.0
virtualenv==15.0.3
